### 2020-10-01 #2715 @whedon
[PRE REVIEW]: ropenblas: Download, Compile and Link OpenBLAS Library with R
**Submitting author:** @prdm0 (<a href="http://orcid.org/0000-0003-1591-8300">Pedro Rafael Marinho</a>)
**Repository:** <a href="https://github.com/prdm0/ropenblas" target ="_blank">https://github.com/prdm0/ropenblas</a>
**Version:** v0.2.8
