### 2020-10-01 #2715 @Kevin-Mattheus-Moerman
@prdm0 thanks for your submission. This paper needs work. It lacks the concise sections we look for, e.g. Statement of need, and is too long. You can likely shorten it significantly by removing content that is already in the readme/documentation.
Please review our guidelines on what the paper should contain: https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain.
You can call `@whedon generate pdf` to update the paper.
