### 2020-10-01 #2715 @Kevin-Mattheus-Moerman
@prdm0 I've flagged this submission for additional scope review by our editorial board which will take about a week. The reason for this scope check is that this submission, to me, at first sight, appears rather minor in terms of functionality and amount of scholarly effort (see our [requirements related to substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort)).
