### 2020-10-01 #2715 @prdm0
Dear @Kevin-Mattheus-Moerman, this package has a lot of effort and has already been referred by important people in the R community. The project is followed by a programmer from the company RStudio and already has more than 6 thousand downloads. The aim of the package is to speed up the R for linear algebra operations. Everything is done within R without relying on GNU / Linux distribution repositories.
