### 2020-10-01 #2715 @prdm0
@Kevin-Mattheus-Moerman The project allows recent versions of R to be installed on any GNU / Linux distribution, even if it does not have the latest version of the R language in the distribution repositories. It also allows us to switch between versions of R.
