### 2020-10-01 #2715 @prdm0
@Kevin-Mattheus-Moerman  The project is very robust and allows to accelerate the R language through the link with the OpenBLAS library. The project is so useful that researchers like Norman Matloff, who wrote several R-language books, featured the project on his Twitter.
