### 2020-10-01 #2715 @prdm0
> @prdm0 I've flagged this submission for additional scope review by our editorial board which will take about a week. The reason for this scope check is that this submission, to me, at first sight, appears rather minor in terms of functionality and amount of scholarly effort (see our [requirements related to substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort)).

@Kevin-Mattheus-Moerman I would like you to be clear when you say that there is no academic effort in a far-fetched R code and that it brings functionalities that underpinned the linear algebra computation of R. Everything is done within R. The package allows you to switch between versions of R and between versions from the OpenBLAS library. In the last month there were more than 800 downloads for a total of more than 6 thousand downloads from the library. The project is followed by people from the field of data science and machine learning and was shared by people who wrote R books and who are highly respected by the community that maintains the language. I have several emails that prove the importance of this package.

This package has a lot to add to JOSS and will certainly bring citations to the Journal.

This package will help any researcher who wishes to perform computationally in algebraic computing, allowing the automatic linking of R to the OpenBLAS library.

Algebraic computations are currently carried out in the most varied research.

I will wait for an editorial board positioning.
