### 2020-10-01 #2715 @prdm0
> Failed to discover a `Statement of need` section in paper

> Failed to discover a `Statement of need` section in paper

The "**Statement of Need**" section has been added.
