### 2020-10-01 #2715 @Kevin-Mattheus-Moerman
@prdm0 thanks for your additional comments. I just want to add I am not implying this work is not useful or of a poor quality. For more on the term "effort" I used see our guidelines: https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort. You'll read there that submissions under 1000 lines of code (judging from the R code contained in this submission) are regularly subjected to additional scope review. I am sure this work will be properly reviewed by our editorial board which includes R specialists (which I am not). Thanks again for your clarifications. You will hear back shortly.
