### 2020-10-01 #2715 @prdm0
@ Kevin-Mattheus-Moerman, I have a question about one of @whedon  output:

```

OK DOIs

- None

MISSING DOIs

- None

INVALID DOIs

- None
```
 is this just a warning that there is no DOI yet? I can disregard it, okay?

Finally, should I wait for the review to reduce the material? What is the minimum number of pages?

Kind regards.
