### 2020-10-01 84053e2
Removing unused reference
