+++ paper/paper.bib
@@ -8,18 +8,6 @@
   url = {https://www.R-project.org/},
 }
 
-@article{ihaka:1996,
-  Author = {Ihaka, Ross and Gentleman, Robert},
-	Journal = {Journal of Computational and Graphical Statistics},
-	Number = 3,
-	Pages = {299--314},
-	Title = {R: A Language for Data Analysis and Graphics},
-	Volume = 5,
-	Year = 1996,
-        url = {https://doi.org/10.1080/10618600.1996.10474713}
-}
-
-
 @misc{gotoblas2,
     Author  = {Kazushige Goto},
     title = {GotoBLAS2 1.13 BSD version},
