### 2020-10-01 db27c7c
Reducing the article from 13 to 9 pages
