+++ paper/paper.md
@@ -58,14 +58,9 @@ The `ropenblas` is a package designed to facilitate the linking of the library O
 The `ropenblas` package is already available on the Comprehensive R Archive Network - CRAN, currently in version 0.2.8 and the project is maintained on GitHub at  <https://github.com/prdm0/ropenblas> where contributors can find others details of the code, information, as well as being able to contribute with the development of the project. Information can also be found on the project website. On the website it is also possible to read the `NEWS.md` file with details of the versions and the focus of the current development. The site is deposited at <https://prdm0.github.io/ropenblas/>. Suggestions for improvements and bug reports can be sent via the link <https://github.com/prdm0/ropenblas/issues>.
 
 
-\begin{figure}[H]
-
-{\centering \includegraphics[width=0.3\linewidth]{logo_ropenblas} 
-
-}
-
-\caption{Computer library logo.}\label{fig:logo}
-\end{figure}
+<!-- ```{r logo, echo = FALSE, message = FALSE, fig.cap = "Computer library logo.", fig.align = "center", out.width="30%"} -->
+<!-- knitr::include_graphics(path = "logo_ropenblas.png") -->
+<!-- ``` -->
 
 
 You can also specify older versions of the OpenBLAS library. Automatically, if no version is specified, the `ropenblas` package will consider the latest version of the library OpenBLAS.
@@ -93,12 +88,12 @@ In addition to dependencies in the form of other packages deposited with CRAN, t
 
 
   1. **GNU Make**: GNU Make utility to maintain groups of programs; 
-	2. **GNU GCC Compiler (C and Fortran)**: The GNU Compiler Collection - C and Fortran frontends.
+  2. **GNU GCC Compiler (C and Fortran)**: The GNU Compiler Collection - C and Fortran frontends.
 
 
-These programs that are described in `SystemRequirements` in the package's `DESCRIPTION` file are essential for compiling the OpenBLAS library and the R language. The functions of the `ropenblas` package are designed to identify the lack of these dependencies external to CRAN, informing the package user which dependencies are missing and suggesting that they should be installed. The other dependencies indexed to CRAN are described in `Imports` in the file  `DESCRIPTION`. These will be installed automatically.
+<!-- These programs that are described in `SystemRequirements` in the package's `DESCRIPTION` file are essential for compiling the OpenBLAS library and the R language. The functions of the `ropenblas` package are designed to identify the lack of these dependencies external to CRAN, informing the package user which dependencies are missing and suggesting that they should be installed. The other dependencies indexed to CRAN are described in `Imports` in the file  `DESCRIPTION`. These will be installed automatically. -->
 
-Other warnings can also be suggested, such as, for example, a problem with the internet connection. All warnings are given very clearly so that the user has no doubts about the problem that may be occurring.
+<!-- Other warnings can also be suggested, such as, for example, a problem with the internet connection. All warnings are given very clearly so that the user has no doubts about the problem that may be occurring. -->
 
 # Installation
 
@@ -122,32 +117,7 @@ The function `last_version_r()` automatically searches, in the official reposito
 
 
 ```r
-> last_version_r(major = 3L)
-```
-
-```
-## $last_version
-## [1] "3.6.3"
-## 
-## $versions
-##  [1] "3.0.0"         "3.0.1"         "3.0.2"         "3.0.3"        
-##  [5] "3.1.0"         "3.1.1"         "3.1.2"         "3.1.3"        
-##  [9] "3.2.0"         "3.2.1"         "3.2.2"         "3.2.3"        
-## [13] "3.2.4-revised" "3.2.4"         "3.2.5"         "3.3.0"        
[…]
