### 2020-10-01 #2715 @prdm0
Paper reduced from 13 to 9 pages.
