### 2020-10-01 #2715 @prdm0
I significantly reduced the number of pages, from 13 to 9.
