### 2020-10-02 #2715 @prdm0
> > @Kevin-Mattheus-Moerman, I have a question about one of @whedon output:
> > ```
> > Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
> >
> > OK DOIs
> >
> > - None
> >
> > MISSING DOIs
> >
> > - None
> >
> > INVALID DOIs
> >
> > - None
> > ```
> >
> >
> > is this just a warning that there is no DOI yet? I can disregard it, okay?
> > Finally, should I wait for the review to reduce the material? What is the minimum number of pages?
> > Kind regards.
>
> In relation to this point_up_2 question. Our system checks if: 1) the references/citations you are using should have DOI's, and 2) if they do if the DOI's given are valid (they link to the correct paper) or not. It looks like you cite 3 "websites" or urls to software or other sites. These likely do not have DOI's so the system finds you currently have no DOI's (and therefore no invalid ones etc). So there is no problem there in principle.
>
> However, on a related note, and getting back to the paper content, it would be good to cite research utilizing this work and also perhaps similar/relevant works/research appearing in the literature. Once you add such references they will need appropriate DOI's in the .bib entries as well. Let me know if you have more questions. For more information check the "What should my paper contain?" section here: https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain

Thanks for the clarification. Understood.
