### 2020-10-02 #2715 @prdm0
> Just a heads up @prdm0 - this paper is still far too long. Most of the detailed documentation of the API should not be in the JOSS paper.
>
> As a guideline, JOSS submissions should be ~1000 words (~4 pages max).

Thank you. I will try to focus now on drastically reducing the paper.
