### 2020-10-02 #2715 @prdm0
Drastic paper reduction from 13 to 4 pages.
