+++ paper/paper.md
@@ -15,7 +15,7 @@ affiliations:
  - name: Department of Statistics, Federal University of Paraíba, João Pessoa, Paraíba - PB, Brazil
    index: 1
 citation_author: Marinho
-date: "2020-10-01"
+date: "2020-10-02"
 year: "2020"
 bibliography: paper.bib
 output: rticles::joss_article
@@ -47,15 +47,15 @@ The `ropenblas` package aims to provide useful and simple experiences to R [@R] 
 
 Linking the R language to the OpenBLAS library can bring several benefits to algebraic computing in R. OpenBLAS is an Open-Source implementation of the Basic Linear Algebra Subprograms - BLAS library that is often the first library option for algebraic computing to be linked in the installation of R on many GNU/Linux distributions. The OpenBLAS library is available at <https://github.com/xianyi/OpenBLAS> and adds optimized implementations of linear algebra kernels that can run optimized on various processor architectures. OpenBLAS is based on the GotoBLAS2 project code in version 1.13 [@gotoblas2], code available under the terms of the BSD license.
 
-The functions of the `ropenblas` package can help the R language user to make this link without leaving the R promprt, as well as allowing the user to choose the version of OpenBLAS to be considered, the last stable version being considered by default . Everything is accomplished by functions without many arguments and without major complications to be used. These functions can be very comforting for users of R on GNU/Linux systems who do not feel safe to run several codes in a terminal that runs Shell Script codes for the language configuration.
+<!-- The functions of the `ropenblas` package can help the R language user to make this link without leaving the R promprt, as well as allowing the user to choose the version of OpenBLAS to be considered, the last stable version being considered by default . Everything is accomplished by functions without many arguments and without major complications to be used. These functions can be very comforting for users of R on GNU/Linux systems who do not feel safe to run several codes in a terminal that runs Shell Script codes for the language configuration. -->
 
-It is very common to see users of R in distributions with repositories not prone to immediate updates of programs using several tutorials found in communities on the internet suggesting several lines of code and steps that can be potentially dangerous or easy to be misunderstood by many users of R language. Being able to compile, enable resources if necessary and switch between versions of R using simple functions to be used and without leaving the command pompt of the R language is attractive.
+<!-- It is very common to see users of R in distributions with repositories not prone to immediate updates of programs using several tutorials found in communities on the internet suggesting several lines of code and steps that can be potentially dangerous or easy to be misunderstood by many users of R language. Being able to compile, enable resources if necessary and switch between versions of R using simple functions to be used and without leaving the command pompt of the R language is attractive. -->
 
-# General package information
+<!-- # General package information -->
 
-The `ropenblas` is a package designed to facilitate the linking of the library OpenBLAS with the language R. The package, which works only for Linux systems, will automatically download the latest source code from the OpenBLAS library and compile the code. The package will automatically bind the language R, through the `ropenblas()` function, to use the OpenBLAS library. Everything will be done automatically regardless of the Linux distribution you are using.
+The `ropenblas` is a package designed to facilitate the linking of the library OpenBLAS with the language R. The package, which works only for Linux systems, will automatically download the latest source code from the OpenBLAS library and compile the code. The package will automatically bind the language R, through the `ropenblas()` function, to use the OpenBLAS library. Everything will be done automatically regardless of the Linux distribution you are using. Enumerating some advantages of the package:
 
[…]
