### 2020-10-02 #2715 @prdm0
Dear @Kevin-Mattheus-Moerman I reduced the paper from 13 to 4 pages, in order to fit with the Journal. Do I need to do anything else or should I wait?
