### 2020-10-12 #2715 @prdm0
> @prdm0 - while this code is somewhat of a wrapper and it is relatively small, after some discussion, we have decided that because of its dynamic properties, it is not just a wrapper, and we will go ahead and review it.

Dear @danielskatz , I'm happy to have my tools being evaluated at JOSS. Enabling distribution-independent GNU/Linux users to link R to optimized versions of BLAS is of great use. Thank you all for understanding the importance of the project.

Kind regards,
prdm0.
