### 2020-10-13 #2715 @timtroendle
👋 @wrathematics, @crew102, @matloff, would any of you be willing to review this submission for JOSS?
