### 2020-10-13 #2715 @matloff
Unfortunately, I have only a passing familiarity with OpenBLAS, simply as a user, and don't feel qualified to review papers on it.

Norm
