### 2020-10-23 #2715 @jgoldfar
Hello Tim; sorry for the delay. I would be happy to help, but will
probably not have time to complete the review well until December.
Best,
Jonathan Goldfarb

* Tim Tröndle <notifications@github.com> [2020-10-18 23:27:31 -0700]:

>👋 @Jgoldfar, @pratikvn, would any of you be willing to review this submission for JOSS?
>
>--
>You are receiving this because you were mentioned.
>Reply to this email directly or view it on GitHub:
>https://github.com/openjournals/joss-reviews/issues/2715#issuecomment-711645562
