### 2020-11-14 prdm0/ropenblas#21 @pratikvn
JOSS review
Hello @prdm0 ,

I am a reviewer for the paper you submitted to JOSS. The package is very useful to a lot of R users and I thank you for creating this package.

I had some comments on the paper. Please find them in the pdf attached. In summary the paper looks good to me. I just had a few suggestions and one aspect that I think is necessary to address:

1. As you claim that the ropenblas provides an improved performance, you should validate this claim by adding some runtime comparisons with and without openblas.

2. Other suggestions are mainly some refactoring and removal of repeated content especially in the second section.

3. There were also some other suggestions mainly addressing some language issues. Some of these might be stylistic, so feel free to disregard them.

[Download the ropenblas_reviewed.pdf](https://github.com/prdm0/ropenblas/files/5541564/ropenblas_reviewed.pdf)
