### 2020-11-15 prdm0/ropenblas#21 @prdm0
Hello @pratikvn , thank you very much for sending us your suggestions for improvements. It will certainly help to improve the quality of the paper. When the improvements are implemented, should I generate a new PDF at [#2769](https://github.com/openjournals/joss-reviews/issues/2769)?

Is there a maximum time limit for improvements to be implemented?

Kind regards.
