### 2020-11-15 #2769 @prdm0
Prezado @timtroendle , Is there a maximum time limit for improvements to be implemented? When the suggestions are implemented, should I generate a new PDF using `@whedon generate pdf`?

Kind regards.
