### 2020-11-20 #2769 @prdm0
Dear @timtroendle, I intend to quickly make the changes suggested by the first reviewer. When they are implemented, should I update the PDF or should I wait for suggestions from the second reviewer?
