### 2020-12-07 prdm0/ropenblas#21 @pratikvn
@prdm0 , I had a few more comments to address some of the Functionality and the Documentation requirements:

1. Can you please re-factor the README.md file and only keep minimal installation instructions, dependency information, a small example and the last section on Advantages in README ? And move the rest to a separate INSTALL.md file. I think this makes the README a little easier to read and if you can add a table of contents, this would make the navigation much better. [See related Stackoverflow question and accepted answer](https://stackoverflow.com/questions/2822089/how-to-link-to-part-of-the-same-document-in-markdown)

2. I guess there are no automated tests you can implement for this repository ? Maybe running on some dataset on the github actions CI or your preferred CI setup ?

3. Could you also add a CONTRIBUTING.md file? This helps other people contribute to this repository. See [github blog](https://github.blog/2012-09-17-contributing-guidelines/).

4. As you claim that the performance is better when using an openblas backend, I think you should verify this. You could even add a plot to the README.
