### 2020-12-07 prdm0/ropenblas#21 @prdm0
@pratikvn  thanks for the suggestions. I will try to implement most of the suggestions.

I am waiting for the second reviewer to present his suggestions so that I can implement everything at once. This would prevent me from modifying the files and documentation that he may already be correcting.

On performance, I will perform some benchmark to prove it, although it is already known that the performance of OpenBLAS is inferior to that of the BLAS library ([see](http://blog.nguyenvq.com/blog/2014/11/10/optimized-r-and-python-standard-blas-vs-atlas-vs-openblas-vs-mkl/)). However, I agree that it would be interesting to add this to the article. This would make the paper easier to read. I had even made benchmarks and removed it, since the Editor had asked to shrink the paper that previously had 13 pages to 4 pages.

There is also a library website that was built at [RMarkdown](https://rmarkdown.rstudio.com/).  Detail in **<https://prdm0.github.io/ropenblas/>**.

I will try to create a REAMD.md with an index to facilitate navigation between the information, following the sections you suggested.
