+++ README.md
@@ -1,235 +1,439 @@
-# ropenblas: Download, Compile and Link OpenBLAS Library with R
+Download, Compile and Link OpenBLAS Library with R
+================
+**Author**: Prof. Dr. Pedro Rafael Diniz Marinho  
+**Package website**: **<https://prdm0.github.io/ropenblas>**  
+  
+  
+  
+**Summary**
+
+-   [1 General aspects](#general-aspects)
+-   [2 Dependencies](#dependencies)
+-   [3 Installation](#installation)
+-   [4 Use](#use)
+    -   [4.1 ‘ropenblas’ function:](#ropenblas-function)
+        -   [4.1.1 Details](#details)
+    -   [4.2 ‘last\_version\_r’ function:](#last_version_r-function)
+    -   [4.3 ‘last\_version\_openblas’
+        function:](#last_version_openblas-function)
+    -   [4.4 ‘rcompiler’ function:](#rcompiler-function)
+    -   [4.5 ‘link\_again’ function:](#link_again-function)
+    -   [4.6 ‘rnews’ function:](#rnews-function)
+-   [5 Advantages of using ropenblas
+    package:](#advantages-of-using-ropenblas-package)
 
 [![last](https://www.r-pkg.org/badges/last-release/ropenblas)](https://CRAN.R-project.org/package=ropenblas)
-[![CRAN_Status_Badge](https://www.r-pkg.org/badges/version/ropenblas)](https://CRAN.R-project.org/package=ropenblas)
+[![CRAN\_Status\_Badge](https://www.r-pkg.org/badges/version/ropenblas)](https://CRAN.R-project.org/package=ropenblas)
 [![total](http://cranlogs.r-pkg.org/badges/grand-total/ropenblas)](https://CRAN.R-project.org/package=ropenblas)
 [![month](https://cranlogs.r-pkg.org/badges/ropenblas)](https://CRAN.R-project.org/package=ropenblas)
 
 <img src="https://raw.githubusercontent.com/prdm0/ropenblas/master/logo.png" height="270" width="250" align="right" />
 
-The [**ropenblas**](https://prdm0.github.io/ropenblas/) is a package designed to facilitate the linking of the library [**OpenBLAS**](https://www.openblas.net/) with the language [**R**](https://www.r-project.org/). The package, which works only for Linux systems, will automatically download the latest source code from the [**OpenBLAS**](https://www.openblas.net/) library and compile the code. The package will automatically bind the language [**R**](https://www.r-project.org/), through the `ropenblas()` function, to use the [**OpenBLAS**](https://www.openblas.net/) library. Everything will be done automatically regardless of the Linux distribution you are using.
-  
-You can also specify older versions of the [**OpenBLAS**](https://www.openblas.net/) library. Automatically, if no version is specified, the [**ropenblas**](https://prdm0.github.io/ropenblas/) package will consider the latest version of the library [**OpenBLAS**](https://www.openblas.net/).
+# 1 General aspects
 
+The [**ropenblas**](https://prdm0.github.io/ropenblas/) is a package
+designed to facilitate the linking of the library
+[**OpenBLAS**](https://www.openblas.net/) with the language
+[**R**](https://www.r-project.org/). The package, which works only for
+Linux systems, will automatically download the latest source code from
+the [**OpenBLAS**](https://www.openblas.net/) library and compile the
+code. The package will automatically bind the language
+[**R**](https://www.r-project.org/), through the `ropenblas()` function,
+to use the [**OpenBLAS**](https://www.openblas.net/) library. Everything
+will be done automatically regardless of the Linux distribution you are
+using.
 
[…]
