+++ README.md
@@ -11,15 +11,15 @@ Download, Compile and Link OpenBLAS Library with R
 -   [2 Dependencies](#2-dependencies)
 -   [3 Installation](#3-installation)
 -   [4 Use](#use)
-    -   [4.1 ‘ropenblas’ function:](#41-ropenblas-function)
-    -   [4.2 ‘last\_version\_r’ function:](#42-last_version_r-function)
+    -   [4.1 ‘ropenblas’ function](#41-ropenblas-function)
+    -   [4.2 ‘last\_version\_r’ function](#42-last_version_r-function)
     -   [4.3 ‘last\_version\_openblas’
         function:](#43-last_version_openblas-function)
-    -   [4.4 ‘rcompiler’ function:](#44-rcompiler-function)
-    -   [4.5 ‘link\_again’ function:](#45-link_again-function)
-    -   [4.6 ‘rnews’ function:](#46-rnews-function)
+    -   [4.4 ‘rcompiler’ function](#44-rcompiler-function)
+    -   [4.5 ‘link\_again’ function](#45-link_again-function)
+    -   [4.6 ‘rnews’ function](#46-rnews-function)
 -   [5 Advantages of using ropenblas
-    package:](#5-advantages-of-using-ropenblas-package)
+    package](#5-advantages-of-using-ropenblas-package)
 
 [![last](https://www.r-pkg.org/badges/last-release/ropenblas)](https://CRAN.R-project.org/package=ropenblas)
 [![CRAN\_Status\_Badge](https://www.r-pkg.org/badges/version/ropenblas)](https://CRAN.R-project.org/package=ropenblas)
