+++ README.Rmd
@@ -1,7 +1,6 @@
 ---
 title: "Download, Compile and Link OpenBLAS Library with R"
 author: |
-  | **Author**: Prof. Dr. Pedro Rafael Diniz Marinho
   | **Package website**: **<https://prdm0.github.io/ropenblas>**
   |
   |
@@ -38,6 +37,28 @@ The ropenblas package, by `rcompiler()` function is also useful if you want to i
 
 The use of the ropenblas package will return warnings that help you proceed with the use of the functions. If your internet is not working or if any dependency on the operating system is not present, the package will let you know. 
 
+# Advantages of using ropenblas package
+
+Some advantages of using the [**ropenblas**](https://prdm0.github.io/ropenblas/) library:
+
+   - Everything is done within the [**R**](https://www.r-project.org/) language;
+
+   - The procedure will be the same for any Linux distribution;
+
+   - The [**OpenBLAS**](https://www.openblas.net/) library will be compiled and you will choose which build version to bind to [**R**](https://www.r-project.org/), regardless of your Linux distribution;
+
+   - If your GNU/Linux distribution does not have updated versions of [**OpenBLAS**](https://www.openblas.net/), it matters little. The ropenblas package fetches the latest stable release of the [**OpenBLAS**](https://www.openblas.net/) library development account on GitHub;
+
+   - You do not need to know Linux well. In some distributions, it may not be so simple for a less experienced user to compile and link the library to the [**OpenBLAS**](https://www.openblas.net/) library with the [**R**](https://www.r-project.org/) language;
+
+   - It is much easier to direct a person to link [**OpenBLAS**](https://www.openblas.net/) with [**R**](https://www.r-project.org/) saying "run `ropenblas()` within [**R**](https://www.r-project.org/)" than asking that person to verify that an unoptimized version of [**BLAS**](http://www.netlib.org/blas/) installed on the system. Then you have to guide the removal of the unoptimized version of [**BLAS**](http://www.netlib.org/blas/) and guide it to the installation of the library [**OpenBLAS**](https://www.openblas.net/) through the most diverse procedures depending on the GNU/Linux distribution used;
+
+   - As stated earlier, the procedure works for any Linux and this includes Android. If your Android is capable of running privileged commands (ROOT) and if you have [**R**](https://www.r-project.org/) installed via Termux with the required dependencies, you can compile and link [**OpenBLAS**](https://www.openblas.net/) with [**R**](https://www.r-project.org/) using ropenblas;
+   
+   - With the `rcompiler()` function you can build any version of [**R**](https://www.r-project.org/) into your computer architecture, which includes the most stable version of the language;
+   
+   - What is the latest stable version of [**R**](https://www.r-project.org/)? Are you too lazy to go to the [**R**](https://www.r-project.org/) site? Run `ropenblas::last_version_r(major = NULL)`.
+
 # Dependencies
 
 You must install the following dependencies on your operating system (Linux):
@@ -72,7 +93,7 @@ The [**ropenblas**](https://prdm0.github.io/ropenblas/) package currently provid
 library(ropenblas)
 ```
 
-## 'ropenblas' function:
+## 'ropenblas' function
 
 Installing, compiling, and linking the [**OpenBLAS**](https://www.openblas.net/) version **0.3.12** library to the [**R**](https://www.r-project.org/) language:
 
@@ -88,7 +109,7 @@ ropenblas(x = "0.3.12")
    
    - If `x = NULL`, the latest stable version of the [**OpenBLAS**](https://www.openblas.net/) library will be compiled and linked to [**R**](https://www.r-project.org/).
   
-### Details
+**Details**
 
[…]
