+++ README.md
@@ -1,24 +1,24 @@
 Download, Compile and Link OpenBLAS Library with R
 ================
-
 **Package website**: **<https://prdm0.github.io/ropenblas>**  
+  
   
   
 **Summary**
 
 -   [1 General aspects](#1-general-aspects)
--   [2 Dependencies](#2-dependencies)
--   [3 Installation](#3-installation)
--   [4 Use](#use)
-    -   [4.1 ‘ropenblas’ function](#41-ropenblas-function)
-    -   [4.2 ‘last\_version\_r’ function](#42-last_version_r-function)
-    -   [4.3 ‘last\_version\_openblas’
-        function:](#43-last_version_openblas-function)
-    -   [4.4 ‘rcompiler’ function](#44-rcompiler-function)
-    -   [4.5 ‘link\_again’ function](#45-link_again-function)
-    -   [4.6 ‘rnews’ function](#46-rnews-function)
--   [5 Advantages of using ropenblas
-    package](#5-advantages-of-using-ropenblas-package)
+-   [2 Advantages of using ropenblas
+    package](#2-advantages-of-using-ropenblas-package)
+-   [3 Dependencies](#3-dependencies)
+-   [4 Installation](#4-installation)
+-   [5 Use](#5-use)
+    -   [5.1 ‘ropenblas’ function](#51-ropenblas-function)
+    -   [5.2 ‘last\_version\_r’ function](#52-last_version_r-function)
+    -   [5.3 ‘last\_version\_openblas’
+        function](#53-last_version_openblas-function)
+    -   [5.4 ‘rcompiler’ function](#53-rcompiler-function)
+    -   [5.5 ‘link\_again’ function](#55-link_again-function)
+    -   [5.6 ‘rnews’ function](#56-rnews-function)
 
 [![last](https://www.r-pkg.org/badges/last-release/ropenblas)](https://CRAN.R-project.org/package=ropenblas)
 [![CRAN\_Status\_Badge](https://www.r-pkg.org/badges/version/ropenblas)](https://CRAN.R-project.org/package=ropenblas)
@@ -82,7 +82,61 @@ proceed with the use of the functions. If your internet is not working
 or if any dependency on the operating system is not present, the package
 will let you know.
 
-# 2 Dependencies
+# 2 Advantages of using ropenblas package
+
+Some advantages of using the
+[**ropenblas**](https://prdm0.github.io/ropenblas/) library:
+
+-   Everything is done within the [**R**](https://www.r-project.org/)
+    language;
+
+-   The procedure will be the same for any Linux distribution;
+
+-   The [**OpenBLAS**](https://www.openblas.net/) library will be
+    compiled and you will choose which build version to bind to
+    [**R**](https://www.r-project.org/), regardless of your Linux
+    distribution;
+
+-   If your GNU/Linux distribution does not have updated versions of
+    [**OpenBLAS**](https://www.openblas.net/), it matters little. The
+    ropenblas package fetches the latest stable release of the
+    [**OpenBLAS**](https://www.openblas.net/) library development
+    account on GitHub;
+
+-   You do not need to know Linux well. In some distributions, it may
+    not be so simple for a less experienced user to compile and link the
+    library to the [**OpenBLAS**](https://www.openblas.net/) library
+    with the [**R**](https://www.r-project.org/) language;
+
+-   It is much easier to direct a person to link
+    [**OpenBLAS**](https://www.openblas.net/) with
+    [**R**](https://www.r-project.org/) saying “run `ropenblas()` within
+    [**R**](https://www.r-project.org/)” than asking that person to
+    verify that an unoptimized version of
+    [**BLAS**](http://www.netlib.org/blas/) installed on the system.
[…]
