+++ README.md
@@ -195,7 +195,7 @@ the [**R**](https://www.r-project.org/) language:
 
 -   [**OpenBLAS**](https://www.openblas.net/) versions tested: 0.3.0,
     0.3.1, 0.3.2, 0.3.3, 0.3.4, 0.3.5, 0.3.6, 0.3.7, 0.3.8, 0.3.9,
-    0.3.10, 0.3.11 and 0.3.12. These are the values that will be passed
+    0.3.10, 0.3.11, 0.3.12 and 0.3.13. These are the values that will be passed
     to `x` in `ropenblas(x)`;
 
 -   If `x = NULL`, the latest stable version of the
