+++ README.md
@@ -292,20 +292,23 @@ The `last_version_openblas()` function automatically searches
 [**OpenBLAS**](https://www.openblas.net/) library versions in the
 official [**GitHub**](https://github.com/xianyi/OpenBLAS) project.
 
-    > last_version_openblas()
-    $last_version
-    [1] "0.3.12"
+```{r}
+> ropenblas::last_version_openblas()
+$last_version
+[1] "0.3.13"
 
-    $versions
-     [1] "0.1alpha1"   "0.1alpha2"   "0.1alpha2.1" "0.1alpha2.2" "0.1alpha2.3" "0.1alpha2.4" "0.1alpha2.5" "0.1.0"      
-     [9] "0.1.1"       "0.2.0"       "0.2.1"       "0.2.2"       "0.2.3"       "0.2.4"       "0.2.5"       "0.2.6"      
-    [17] "0.2.7"       "0.2.8"       "0.2.9"       "0.2.9.rc1"   "0.2.9.rc2"   "0.2.10"      "0.2.10.rc1"  "0.2.10.rc2" 
-    [25] "0.2.11"      "0.2.12"      "0.2.13"      "0.2.14"      "0.2.15"      "0.2.16"      "0.2.16.rc1"  "0.2.17"     
-    [33] "0.2.18"      "0.2.19"      "0.2.20"      "0.3.0"       "0.3.1"       "0.3.2"       "0.3.3"       "0.3.4"      
-    [41] "0.3.5"       "0.3.6"       "0.3.7"       "0.3.8"       "0.3.9"       "0.3.10"      "0.3.11"      "0.3.12"     
+$versions
+ [1] "0.1alpha1"   "0.1alpha2"   "0.1alpha2.1" "0.1alpha2.2" "0.1alpha2.3" "0.1alpha2.4" "0.1alpha2.5"
+ [8] "0.1.0"       "0.1.1"       "0.2.0"       "0.2.1"       "0.2.2"       "0.2.3"       "0.2.4"      
+[15] "0.2.5"       "0.2.6"       "0.2.7"       "0.2.8"       "0.2.9"       "0.2.9.rc1"   "0.2.9.rc2"  
+[22] "0.2.10"      "0.2.10.rc1"  "0.2.10.rc2"  "0.2.11"      "0.2.12"      "0.2.13"      "0.2.14"     
+[29] "0.2.15"      "0.2.16"      "0.2.16.rc1"  "0.2.17"      "0.2.18"      "0.2.19"      "0.2.20"     
+[36] "0.3.0"       "0.3.1"       "0.3.2"       "0.3.3"       "0.3.4"       "0.3.5"       "0.3.6"      
+[43] "0.3.7"       "0.3.8"       "0.3.9"       "0.3.10"      "0.3.11"      "0.3.12"      "0.3.13"     
 
-    $n
-    [1] 48
+$n
+[1] 49
+```
 
 ## 5.4 ‘rcompiler’ function
 
