### 2020-12-15 43971f9
Update README.md. Adding animated examples via the asciicast package. Updating site.
