+++ inst/asciicast/example_asciicast_last_version_openblas.R
@@ -1,0 +1,7 @@
+#' Title: Using the last_version_openblas() function
+#' Rows: 24
+
+# The last_version_openblas() function automatically searches
+# OpenBLAS library versions in the official GitHub project.
+
+ropenblas::last_version_openblas()
