+++ inst/asciicast/generating svg files.R
@@ -1,0 +1,23 @@
+cast <- asciicast::record("~/Downloads/ropenblas/inst/asciicast/example_asciicast_last_version_r.R")
+asciicast::write_svg(cast, path = "~/Downloads/ropenblas/inst/asciicast/last_version_r.svg",
+                     window = TRUE, cursor = TRUE)
+
+cast <- asciicast::record("~/Downloads/ropenblas/inst/asciicast/example_asciicast_last_version_openblas.R")
+asciicast::write_svg(cast, path = "~/Downloads/ropenblas/inst/asciicast/last_version_openblas.svg", window = TRUE,
+                     cursor = TRUE)
+
+cast <- asciicast::record("~/Downloads/ropenblas/inst/asciicast/example_asciicast_link_again.R")
+asciicast::write_svg(cast, path = "~/Downloads/ropenblas/inst/asciicast/link_again.svg", window = TRUE,
+                     cursor = TRUE)
+
+cast <- asciicast::record("~/Downloads/ropenblas/inst/asciicast/example_asciicast_ropenblas.R")
+asciicast::write_svg(cast, path = "~/Downloads/ropenblas/inst/asciicast/ropenblas.svg", window = TRUE,
+                     cursor = TRUE)
+
+cast <- asciicast::record("~/Downloads/ropenblas/inst/asciicast/example_asciicast_rcompiler.R")
+asciicast::write_svg(cast, path = "~/Downloads/ropenblas/inst/asciicast/rcompiler.svg", window = TRUE,
+                     cursor = TRUE)
+
+cast <- asciicast::record("~/Downloads/ropenblas/inst/asciicast/example_asciicast_install.R")
+asciicast::write_svg(cast, path = "~/Downloads/ropenblas/inst/asciicast/install.svg", window = TRUE,
+                     cursor = TRUE)
