+++ README.md
@@ -120,12 +120,7 @@ Some advantages of using the
     [**OpenBLAS**](https://www.openblas.net/) through the most diverse
     procedures depending on the GNU/Linux distribution used;
 
--   As stated earlier, the procedure works for any Linux and this
-    includes Android. If your Android is capable of running privileged
-    commands (ROOT) and if you have [**R**](https://www.r-project.org/)
-    installed via Termux with the required dependencies, you can compile
-    and link [**OpenBLAS**](https://www.openblas.net/) with
-    [**R**](https://www.r-project.org/) using ropenblas;
+<!-- - As stated earlier, the procedure works for any Linux and this includes Android. If your Android is capable of running privileged commands (ROOT) and if you have [**R**](https://www.r-project.org/) installed via Termux with the required dependencies, you can compile and link [**OpenBLAS**](https://www.openblas.net/) with [**R**](https://www.r-project.org/) using ropenblas; -->
 
 -   With the `rcompiler()` function you can build any version of
     [**R**](https://www.r-project.org/) into your computer architecture,
@@ -157,20 +152,13 @@ package. This will allow you to install the
 [**ropenblas**](https://prdm0.github.io/ropenblas/) package directly
 from GitHub. To install, after installing the **devtools** package, do:
 
-``` r 
-remotes::install_github(repo = "prdm0/ropenblas", force = TRUE)
-```  
+    remotes::install_github(repo = "prdm0/ropenblas", force = TRUE)
 
 or
 
-```r    
-install.packages("ropenblas")
-```
+    install.packages("ropenblas")
 
-
-**Note**: If you want to access the latest features of the
-[**ropenblas**](https://prdm0.github.io/ropenblas/) package, install it
-using the first procedure.
+<!-- **Note**: If you want to access the latest features of the [**ropenblas**](https://prdm0.github.io/ropenblas/) package, install it using the first procedure. -->
 
 # 5 Use
 
@@ -187,90 +175,22 @@ Installing, compiling, and linking the
 [**OpenBLAS**](https://www.openblas.net/) version **0.3.13** library to
 the [**R**](https://www.r-project.org/) language:
 
-```r
-> ropenblas::ropenblas("0.3.13")
-cloning into '/tmp/openblas'...
-Receiving objects:   1% (452/45124),  246 kb
-Receiving objects:  11% (4964/45124), 1930 kb
-Receiving objects:  21% (9477/45124), 6813 kb
-Receiving objects:  31% (13989/45124), 8549 kb
-Receiving objects:  41% (18501/45124), 13484 kb
-Receiving objects:  51% (23014/45124), 16443 kb
-Receiving objects:  61% (27526/45124), 19627 kb
-Receiving objects:  71% (32039/45124), 22779 kb
-Receiving objects:  81% (36551/45124), 24698 kb
-Receiving objects:  91% (41063/45124), 27530 kb
-Receiving objects: 100% (45124/45124), 28694 kb, done.
+<img src="inst/asciicast/ropenblas.svg" style="display: block; margin: auto;" />
 
-● The latest version of OpenBLAS is already in use. Do you want to compile and link again? (yes/no): no
-```
-**Notes**:
-
--   You do not have to in every section of
-    [**R**](https://www.r-project.org/) make use of the `ropenblas()`
-    function. Once the function is used,
-    [**R**](https://www.r-project.org/) will always consider using the
-    [**OpenBLAS**](https://www.openblas.net/) library in future
-    sections.
-
--   [**OpenBLAS**](https://www.openblas.net/) versions tested: 0.3.0,
[…]
