### 2020-12-16 0c2ca41
Add new logo
