### 2020-12-16 c2177c2
Add new logo
