### 2020-12-16 prdm0/ropenblas#21 @prdm0
@pratikvn, I made several modifications to [README.md](https://github.com/prdm0/ropenblas/blob/master/README.md) and others requested. I believe that it is now easier to read [README.md](https://github.com/prdm0/ropenblas/blob/master/README.md). I put the index with the navigation links and created the file [CONTRIBUTING.md](https://github.com/prdm0/ropenblas/blob/master/CONTRIBUTING.md).
