### 2020-12-16 ec330a9
Updating directories, packages and minor changes
