### 2020-12-16 fddfea2
Update size logo
