+++ README.Rmd
@@ -16,14 +16,16 @@ toccolor: 'blue'
 
 
 **ropenblas** 
-<a href='logo.png'><img src='logo.png' align="right" height="200" /></a> 
 
 [![last](https://www.r-pkg.org/badges/last-release/ropenblas)](https://CRAN.R-project.org/package=ropenblas)
 [![CRAN_Status_Badge](https://www.r-pkg.org/badges/version/ropenblas)](https://CRAN.R-project.org/package=ropenblas)
 [![total](http://cranlogs.r-pkg.org/badges/grand-total/ropenblas)](https://CRAN.R-project.org/package=ropenblas)
 [![month](https://cranlogs.r-pkg.org/badges/ropenblas)](https://CRAN.R-project.org/package=ropenblas)
 
-<br/><br/><br/><br/><br/>
+```{r, fig.align='center', echo=FALSE, out.width=219}
+knitr::include_graphics("https://raw.githubusercontent.com/prdm0/ropenblas/master/logo.png")
+```
+
 
 <!-- <img src="https://raw.githubusercontent.com/prdm0/ropenblas/master/logo.png" height="270" width="250" align="right" /> -->
 
