### 2021-01-05 #2769 @timtroendle
@openjournals/dev, sorry for the trouble but Whedon did not assign the new reviewer to this issue and I am unable to assign them manually. Is there anything I can do here?
