### 2021-01-05 #2769 @arfon
> @openjournals/dev, sorry for the trouble but Whedon did not assign the new reviewer to this issue and I am unable to assign them manually. Is there anything I can do here?

@timtroendle - I'm not sure exactly who has the issue here but @sahilseth should be able to complete their review (update checkboxes etc.) fine, @jgoldfar needs to accept the invite in the message above before they will be able to update their checkboxes.

@timtroendle - just as an FYI - adding reviewers after the pre-review stage is generally a little more challenging (as you're finding out!) which is why we prefer to have the reviewers assigned in the pre-review issue if at all possible.
