### 2021-01-22 #2769 @prdm0
I received @pratikvn reviews on the PROBLEM of the ropenblas project <https://github.com/prdm0/ropenblas/issues/21>. @jgoldfar will also send in the same way or will something here also be sent?
