+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: ropenblas
 Type: Package
 Title: Download, Compile and Link 'OpenBLAS' Library with R
-Version: 0.2.8
+Version: 0.2.9
 Authors@R: 
     person(given = "Pedro Rafael",
            family = "D. Marinho",
