### 2021-02-18 #2769 @myousefi2016
Hi @timtroendle, thanks for giving me this opportunity to review [ropenblas](https://github.com/prdm0/ropenblas). One quick comment for @prdm0 to speed up the review process a bit: The software paper is great and describes very interesting and useful details. I just reflected that in my evaluation in the **Software paper** section. But, for _Quality of writing_: I suggest to double check with an native English speaker (I'm not a native English speaker either) to make sure there is no typo or grammatical error to improve the level of writing quality.
