### 2021-02-18 #2769 @prdm0
Dear @myousefi2016 , thanks for the suggestion. I will revise the text as suggested. I look forward to the other suggestions, if any, in [**issues**](https://github.com/prdm0/ropenblas/issues) of the package project in the Github. Kind regards.
