### 2021-03-08 #2769 @prdm0
Dear Editors,

I proceeded with the changes suggested by the reviewers. I modified the documentation on GitHub and the library website as suggested by one of the reviewers, as well as made several changes to the paper. Among them I mention the inclusion of a benchmark, changes in writing and a better explanation of some concepts.
