### 2021-03-09 prdm0/ropenblas#21 @pratikvn
@prdm0 , thanks for adding the changes. As one of the final steps, I would strongly recommend to pass the whole paper through a grammar correctness check. You can use free tools such as Grammarly (https://www.grammarly.com/) to improve your sentence structure and the overall readability of your paper.
