### 2021-03-09 prdm0/ropenblas#21 @prdm0
Hello @pratikvn, I will do that. Thanks for the sugestion.
