### 2021-03-09 prdm0/ropenblas#21 @prdm0
Dear @pratikvn , I followed your suggestion and made some corrections to the text using the tool. Until the text behaved very well and few errors were found. Thank you very much for the suggestion
