### 2021-03-10 #2769 @timtroendle
Thank you for the update, @prdm0. @pratikvn and @myousefi2016, could you please have a look at the revisions and update your corresponding checklist above? Also, a review issue is still open: https://github.com/prdm0/ropenblas/issues/21. @pratikvn, can you please close the issue if the changes have been implemented sufficiently?
