### 2021-03-10 prdm0/ropenblas#21 @pratikvn
@prdm0 , some final comments:
[annotated.pdf](https://github.com/prdm0/ropenblas/files/6114062/annotated.pdf)
