### 2021-03-10 #2769 @pratikvn
@timtroendle , Everything looks good to me modulo some final paper changes which I have provided on https://github.com/prdm0/ropenblas/issues/21 .

After these changes have been implemented, I can close that issue and we can proceed here.
