### 2021-03-10 prdm0/ropenblas#21 @prdm0
@pratikvn, thanks for the suggestions. I will provide these improvements.
