### 2021-03-10 #2769 @prdm0
@pratikvn, obrigado pelas sugestões. Irei proceder com as melhorias e em breve gerarei um novo PDF.
