### 2021-03-11 #2769 @prdm0
Dear @pratikvn and @timtroendle, I made all the requested changes. @pratikvn I made the corrections to the text you requested, I eliminated unnecessary parts, as indicated in [prdm0/ropenblas#21](https://github.com/prdm0/ropenbla/issues/21). I also improved the benchmarks for building a presentable table. The benchmark plot now includes two algebraic routines for comparison. The new PDF of the paper can be found in the comment below.
