### 2021-03-11 prdm0/ropenblas#21 @pratikvn
The paper looks good from my side.
