### 2021-03-16 #2769 @prdm0
Dear @timtroendle, how long does it take for the paper to be listed on the JOSS page? Best regards.
