### 2021-03-18 #2769 @prdm0
Dear @timtroendle ,

1 - <https://github.com/prdm0/ropenblas/tree/v0.2.9>, **tag v0.2.9**;
2 - <https://zenodo.org/record/4618251#.YFOGiytv83w>;
3 - Done metadata;
4 - <https://doi.org/10.5281/zenodo.4618251>, **DOI**: [10.5281/zenodo.4618251](https://doi.org/10.5281/zenodo.4618251).
