### 2021-03-19 #2769 @timtroendle
Dear @prdm0, thanks for this! Your submission is almost ready for publication. However, the paper needs more work. With more than 2000 words and 6 pages, it’s too long. Can you please shorten the paper to about 1000 words or less? Here are two possibilities:

First, the paper contains a lot of information taken from the documentation. This information is not required in the JOSS article. This concerns in particular the installation instructions and the function overview in the “Brief explanation” section. You could shorten this section drastically be limiting it to the most important function(s), or by discussing an example use only.

Second, the paper contains redundancy: the sections “Summary”, “Statement of Need”, and “Introduction” have quite some overlap. If you streamline these sections, you should be able to cut the number of words significantly. You could, for example, discuss the importance of computational efficiency and OpenBLAS in the Summary section; explain the benefits of ropenblas in the “Statement of Need” section; and reveal further details of ropenblas in a third section (likely not called “Introduction”). This is only a suggestion and I leave the details up to you.
