### 2021-03-19 #2769 @prdm0
Dear @timtroendle, I made the reductions so that there was a reduction of one page. I believe that it complies with the other papers in JOSS that have up to a maximum of 5 pages.
