### 2021-03-19 #2769 @prdm0
Dear @timtroendle, it decreases even more and the paper, The material now has less than 2000 words, more precisely with 1998.

```
[prdm0@samsung paper]$ pdftotext paper.pdf - | wc -w
1998
```
