### 2021-03-22 #2769 @prdm0
@timtroendle, when possible, let me know according to the program you use to count words if the PDF above is suitable. I made a comparison with the latest publications, considering my way of telling, the paper has fewer words than those last publications.

I reduced it by a little more than one page. I believe that the paper is now suitable.

Best regards.
