### 2021-03-22 #2769 @timtroendle
Thank you for the offer @prdm0, but this is not about the exact number of words. Rather, it's about having the right content in the article. Let me have a look at your most recent changes and come back to you soon.
