### 2021-03-22 #2769 @timtroendle
@prdm0, I made some edits to your draft (attached). Can you please work these edits in?
[10.21105.joss.02769-4-TT.pdf](https://github.com/openjournals/joss-reviews/files/6182863/10.21105.joss.02769-4-TT.pdf)

I want to highlight the "Brief explanation" section here. As I mentioned above, large parts of this section are taken from the documentation and provide a level of detail that is not necessary for a JOSS article. Can you please remove or shorten this section? There is no need to discuss individual functions of ropenblas in this article.
