### 2021-03-22 #2769 @prdm0
I will remove all explanations of the functions and leave only the benchmark.
