### 2021-03-22 #2769 @prdm0
@timtroendle I made all the modifications. The PDF with the modified paper is below.
