+++ paper/paper.md
@@ -128,7 +128,7 @@ Through a benchmark it is possible to better understand the performance gain tha
 
 \begin{figure}[H]
 
-{\centering \includegraphics[width=0.6\linewidth]{benchmark} 
+{\centering \includegraphics[width=0.5\linewidth]{benchmark} 
 
 }
 
