+++ paper/paper.Rmd
@@ -123,7 +123,7 @@ Table 1 below presents the benefit of considering an optimized version of BLAS. 
 
 Through a benchmark it is possible to better understand the performance gain that can be achieved by linking the R language to the OpenBLAS library. Figure 2 presents the benchmarks in the form of a violin plot, in which 100 reproductions of the `svd(X)` expression were considered, in the form of the code above, with the R linked to the BLAS library and linked to the OpenBLAS library, respectively, on the same hardware. It was observed that the average time of execution of the routine `svd(X)` considering the OpenBLAS library was less than 10 times the time necessary to execute it in R linking to a non-optimized version of BLAS, being the average time of 0.64 and 6.52 seconds, respectively.
 
-```{r bench, echo = FALSE, fig.cap = "Benchmarks of a decomposition of singular and inverse value of a matrix of dimension 1000 x 1000.", fig.align = "center", out.width="60%"}
+```{r bench, echo = FALSE, fig.cap = "Benchmarks of a decomposition of singular and inverse value of a matrix of dimension 1000 x 1000.", fig.align = "center", out.width="50%"}
 knitr::include_graphics(path = "benchmark.png") 
 ```
 
