### 2021-03-22 #2769 @myousefi2016
Hi @prdm0, just a quick comment here: please finalize all your changes to the manuscript before generating pdf. It's not necessary to do some incremental changes and then generate the pdf. I see there are lots of generate pdf commands here which makes this issue a bit lengthy and hard to follow. It is possible to avoid this by finalizing all your changes and just have one generate pdf command at the end.
