### 2021-03-22 #2769 @prdm0
> Hi @prdm0, just a quick comment here: please finalize all your changes to the manuscript before generating pdf. It's not necessary to do some incremental changes and then generate the pdf. I see there are lots of generate pdf commands here which makes this issue a bit lengthy and hard to follow. It is possible to avoid this by finalizing all your changes and just have one generate pdf command at the end.

Dear @myousefi2016 I know that. However, the PDF was in a format here in RStudio and when the graphic was generated it was on another page. That's why it was generated. But don't worry, I know that and the last generation has been realized and is above your comment.
