+++ _pkgdown.yml
@@ -1,0 +1,4 @@
+destination: docs/
+template:
+  params:
+    bootswatch: flatly # https://bootswatch.com/flatly/
