### 2021-03-23 #2769 @prdm0
> I'm sorry @prdm0, I'm afraid I can't do that. That's something only editor-in-chiefs are allowed to do.

@timtroendle , Pardon me. I thought that I should run the command since the tag had changed to recommended paper for acceptance. I believe that now nothing else depends on me. If you need anything, I'm at your disposal.
