### 2021-03-30 #2769 @kthyng
@openjournals/dev Two names of the author are missing in the XML file — does this mean the paper won't have the correct author name?
