### 2021-03-30 #2769 @kthyng
@prdm0 I am looking through your paper and it needs quite a bit of editing. I started working on it in this [PR](https://github.com/prdm0/ropenblas/pull/28) but I changed sentences substantially to try to reflect my understanding of your writing. Please make sure I have done so accurately before merging.

Quite a bit more editing is necessary. I couldn't understand line 42 in the .md file to try to edit it. There is also a lot of repetition. For example, the fact that the Linux distribution doesn't matter and that behavior in the package is the same was mentioned at least 3 times in the first 1.5 pages. Please read through the manuscript carefully and edit it. The recommended work count for the paper is 250-1000 words so no need to add anything that isn't necessary.
