### 2021-03-30 #2769 @prdm0
> > @openjournals/dev Two names of the author are missing in the XML file — does this mean the paper won't have the correct author name?
>
> @prdm0 - what should your `surname` be here? https://github.com/openjournals/joss-papers/pull/2158/files#diff-93ddcb6dc29bed45dbfd3d925b04a9e6f5f6932c1205907d7e70ab16fe204d20R40

Submeti as modificações.

Name: Pedro Rafael
Last name: Diniz Marinho
