### 2021-03-30 092112c
Update paper. Accepting suggestions from Avatar Kristen Thyng.
